Fix the androideabi toolchain not matching host
The included method of determining the host platform in the androideabi
would not match to windows if RUBY_PLATFORM returned mingw, wince, or
emx. As an added pre-caution, using a case statement with regexp
matching will better determine the host platform since it won't be case
sensitive and for windows is set to match more specifially than just
"win".