replace is_minor_gc() to is_generational() in mrb_incremental_gc()
in the current code, the only chance to trigger the major GC is
out of memory in mrb_realloc_simple(). it would be good to trigger
a major GC in mrb_incremental_gc() if nessecary, for the fact that
the only difference between major GC and minor GC is the difference
on white objects, we can keep the logic same.