Support override of + - * / for Fixnum and Float.
For support of mruby-bignum (https://github.com/chasonr/mruby-bignum/).

A set of flags is added to mrb_state, corresponding to each of eight numeric
methods with existing optimizations:  + - * / for Fixnum and for Float.
If the flag is set, the method has not been overriden and the optimization
remains in effect.  If clear, the use of the operator produces a method call.
Thus these methods can be overriden, just like any other.