Regularized #include under include/mruby/*.h according to https://gcc.gnu.org/onlinedocs/cpp/Include-Syntax.html
According to C Preprocessor syntax, the #include "file" syntax first searches in
the directory containing the current file, then in the quote directories
and then the same directories used for <file>.
The original call, e.g. #include "mruby/common.h" from within the mruby folder,
requires to have -Ipath/to/mruby/include/mruby in compilation (which is not
always the case).