Reject next, break and return with blocks.
Statements such as `next [] do; end` cause a segfault on latest master.
This is due to the fact that it is parsed as a block_call, yet
call_with_block cannot handle these nodes. `yield` with a block was
already rejected, so I added a bunch of similar checks for `next`,
`break` and `return`.