Fix segfault when removing ivars on immediates (fixes #4519)
The problem is that `mrb_iv_frozen` treats the `mrb_value` as a pointer,
but it can also be an immediate (`nil`, a fixnum, etc). This leads to
a segfault since the immediate's value is not a valid pointer.

Looking at `mrb_obj_frozen` and comparing with CRuby behaviour,
immediates should always be treated as frozen. This commit checks
whether the value involved is an immediate and raises `FrozenError` if
so.

Also changes signature of `mrb_frozen_error` so that it can be used with
`mrb_value`s as suggested by shuujii.