Retrieve test for `%l` of `mrb_vformat()`; ref #4613
The following compilation error occurs on Visual Studio 2017 in AppVeyor if
commented out part in previous `mrbgems/mruby-test/vformat.c` is enabled.

  CC    mrbgems/mruby-test/vformat.c -> build/cxx_abi/mrbgems/mruby-test/vformat.obj
  C:\projects\mruby\mrbgems\mruby-test\vformat.c(176): fatal error C1060: compiler is out of heap space
  INTERNAL COMPILER ERROR in 'C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\VC\Tools\MSVC\14.16.27023\bin\HostX64\x64\cl.exe'
      Please choose the Technical Support command on the Visual C++
      Help menu, or open the Technical Support help file for more information

This error seems to occur only C++, and not occur Visual Studio 2013/2015.

The cause seems to be deep nested conditional operator (`? :`). So I decided
to generate `mrbgems/mruby-test/vformat.c` with ERB to avoid it.