Prohibit changes by `RARRAY_PTR()` and `ARY_PTR()`
Previously, explicit write-barriers were required to change content using these macros.

From now on it can use `ARY_PTR_NEED_WB()` instead if it wants to get a pointer without `const` that requires write-barrier.
But if it wants to change only a few contents of the array object, the `mrb_ary_set()` function will be more useful.

With this change, the following function specifications will change:

- The `mrb_get_argv()` function will now return `const mrb_value *`.
  This is because it is difficult for the caller to check if it is a splat argument (array object) and to write-barrier if necessary.
- The "a"/"*" specifier of the `mrb_get_args()` function will now return `const mrb_value *`.
  This is because it is difficult for the caller to check if it is an array object and write-barrier if necessary.
  And it requires calling `mrb_ary_modify()` on the unmodified array object, which is also difficult (this is similar to #5087).