Generate `LICENSE.yml` file
Outputs the `<build-dir>/LICENSE.yml` file by combining the `mruby/LICENSE` and the files related to the license terms for mrbgems added to the `build_config` file.
If it only wants to generate a `LICENSE.yml` file, run `rake license`.

Many common license terms are required to include themselves.  Therefore, the purpose is to reduce the effort of users to put them together.
Note, however, that it is ideal and the generated `LICENSE.yml` file is probably incomplete.

If there is a license terms file in the top directory of mrbgem, it will be automatically collected.
Therefore, it is now recommended to put the `LICENSE` file by the mrbgem author in each top directory.
The target to be imported as a license clause file is one of the combinations of `["LICENSE","COPYRIGHT","COPYING"].product(["",".txt",".md"])`.
Priority: `LICENSE` > `LICENSE.txt` > `LICENSE.md` > ... > `COPYING.txt` > `COPYING.md`

In addition, the user can optionally change or add the file name through the `MRuby::Gem::Specification#terms` property.
If it specifies a relative path, the base directory is the top directory of each mrbgem.

```ruby
MRuby::Gem::Specification.new("my-gem") do |spec|
  spec.terms << "OPTIONAL-LICENSE-FILE"
  spec.terms << "path/to/ANOTHER-LICENSE-FILE"
  ...
end
```

The mrbgems that do not have a `LICENSE` file and are not set to` spec.terms` are now warned when `rake` is done.

The APIs added for users are:
- `lib/mruby/build.rb`
  - `MRuby::Build#terms` property for the build configuration
- `lib/mruby/gem.rb`
  - `MRuby::Gem::Specification#terms` property for the GEM author